% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{myApp}
\alias{myApp}
\title{Iniciar la aplicación Shiny}
\usage{
myApp()
}
\value{
No retorna ningún valor, pero inicia la aplicación Shiny.
}
\description{
Esta función inicializa la aplicación Shiny cargando las interfaces de usuario (UI) y
los servidores desde archivos separados, y luego inicia la aplicación Shiny.
}
\examples{
myApp()

}
\seealso{
\code{\link{shinyApp}}
}
\keyword{internal}
