% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_helpers.R
\name{renderBarPlot}
\alias{renderBarPlot}
\title{Renderiza un gráfico de barras utilizando ggplot2 en el UI de Shiny.}
\usage{
renderBarPlot(
  data,
  x,
  y,
  fill,
  title,
  xlab,
  ylab,
  fillLab = fill,
  facet = FALSE,
  facet_var = NULL
)
}
\arguments{
\item{data}{Conjunto de datos a utilizar en el gráfico.}

\item{x}{Variable para el eje x.}

\item{y}{Variable para el eje y.}

\item{fill}{Variable para asignar colores de relleno a las barras.}

\item{title}{Título del gráfico.}

\item{xlab}{Etiqueta del eje x.}

\item{ylab}{Etiqueta del eje y.}

\item{fillLab}{Etiqueta para la leyenda de colores de relleno.}

\item{facet}{Booleano que indica si se deben agregar facetas al gráfico.}

\item{facet_var}{Variable para facetas (si \code{facet} es TRUE).}
}
\value{
No retorna un valor, pero imprime el gráfico generado en la consola.
}
\description{
Esta función toma un conjunto de datos \code{data} y diversos parámetros para generar
un gráfico de barras con ggplot2. Puede incluir opciones adicionales como facetas.
}
\examples{
# Ejemplo de uso:
renderBarPlot(homiEdad, "x", "y", "fill", "Title", "X Label", "Y Label")

}
