% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_helpers.R
\name{renderDataTable}
\alias{renderDataTable}
\title{Renderiza una tabla de datos utilizando el paquete DT en el UI de Shiny.}
\usage{
renderDataTable(filtered_data)
}
\arguments{
\item{filtered_data}{Conjunto de datos a utilizar en la tabla.}
}
\value{
No retorna un valor, pero renderiza la tabla en el UI de Shiny.
}
\description{
Esta función toma un conjunto de datos \code{filtered_data} y utiliza la librería DT
para renderizar una tabla interactiva en el UI de Shiny.
}
\examples{
# Ejemplo de uso:
renderDataTable(filtered_data)

}
