% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_helpers.R
\name{renderLinePlot}
\alias{renderLinePlot}
\title{Renderiza un gráfico de línea utilizando ggplot2 en el UI de Shiny.}
\usage{
renderLinePlot(
  data,
  x,
  y,
  group,
  color,
  title,
  xlab,
  ylab,
  colorlab = color,
  facet = FALSE,
  facet_var = NULL
)
}
\arguments{
\item{data}{Conjunto de datos a utilizar en el gráfico.}

\item{x}{Variable para el eje x.}

\item{y}{Variable para el eje y.}

\item{group}{Variable para agrupar los datos.}

\item{color}{Variable para asignar colores.}

\item{title}{Título del gráfico.}

\item{xlab}{Etiqueta del eje x.}

\item{ylab}{Etiqueta del eje y.}

\item{colorlab}{Etiqueta para la leyenda de colores.}

\item{facet}{Booleano que indica si se deben agregar facetas al gráfico.}

\item{facet_var}{Variable para facetas (si \code{facet} es TRUE).}
}
\value{
No retorna un valor, pero imprime el gráfico generado en la consola.
}
\description{
Esta función toma un conjunto de datos \code{data} y diversos parámetros para generar
un gráfico de línea con ggplot2. Puede incluir opciones adicionales como facetas.
}
\examples{
# Ejemplo de uso:
renderLinePlot(homiEdad, "x", "y", "group", "color", "Title", "X Label", "Y Label")

}
