% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R, R/server_helpers.R
\name{renderMap}
\alias{renderMap}
\title{Renderiza un mapa utilizando ggplot2 en el UI de Shiny.}
\usage{
renderMap(data, fill, title, fill_lab = fill, fill_color = "blue")

renderMap(data, fill, title, fill_lab = fill, fill_color = "blue")
}
\arguments{
\item{data}{Conjunto de datos a utilizar en el mapa. Debe ser un objeto que contenga datos espaciales,
como un objeto de clase 'sf' de la librería sf.}

\item{fill}{Variable para la escala de colores. Debe ser una columna del conjunto de datos \code{data} que
contiene los valores numéricos que se utilizarán para determinar la intensidad del color en el mapa.}

\item{title}{Título del mapa. Debe ser una cadena de caracteres que describa el contenido del mapa.}

\item{fill_lab}{Etiqueta para la leyenda de colores. Debe ser una cadena de caracteres que describa
la variable representada por la escala de colores.}

\item{fill_color}{Color base para la escala de colores. Debe ser un nombre de color válido en R o un
código hexadecimal que represente un color válido.}
}
\value{
Esta función no retorna un valor, pero imprime el mapa generado en la consola.

Esta función no retorna un valor, pero imprime el mapa generado en la consola.
}
\description{
Esta función toma un conjunto de datos y diversos parámetros para generar un mapa utilizando ggplot2.

Esta función toma un conjunto de datos y diversos parámetros para generar un mapa utilizando ggplot2.
}
\examples{
# Ejemplo de uso:
renderMap(datos_delito_seleccionado, "Casos", "Incidencia de casos de Delito X")

# Ejemplo de uso:
renderMap(datos_delito_seleccionado, "Casos", "Incidencia de casos de Delito X")

}
