% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_helpers.R
\name{updateCheckboxGroup}
\alias{updateCheckboxGroup}
\title{Actualiza un grupo de checkbox en el UI de Shiny.}
\usage{
updateCheckboxGroup(session, inputId, input, data)
}
\arguments{
\item{session}{Una sesión de Shiny.}

\item{inputId}{Identificación única del grupo de checkbox en el UI.}

\item{input}{Objeto \code{input} de Shiny que contiene los valores actuales del UI.}

\item{data}{Conjunto de datos que proporciona las opciones para el grupo de checkbox.}
}
\value{
No retorna un valor, pero actualiza el UI de Shiny según la lógica descrita.
}
\description{
Esta función toma como entrada el \code{session}, el \code{inputId} correspondiente al grupo
de checkbox, el objeto \code{input} y el conjunto de datos \code{data}. Dependiendo de si el
grupo de checkbox está vacío o no, la función llena o vacía el grupo respectivamente.
}
\examples{
# Ejemplo de uso:
updateCheckboxGroup(session, "checkGroup_snmv", input, homiEdad)

}
